+++ src/lib/common/models/Call.ts
@@ -366,7 +366,7 @@ class Call {
                     if (callproperties.find((el) => { return val == el ;})) {
                         that[val] = data[val];
                     } else {
-                        console.log("WARNING : One property of the parameter of updateCall method is not present in the Call class can not update Call with : ", val);
+                        //console.log("WARNING : One property of the parameter of updateCall method is not present in the Call class can not update Call with : ", val);
                     }
                 });
         }
@@ -391,7 +391,7 @@ class Call {
                 Object.getOwnPropertyNames(data).forEach(
                     (val, idx, array) => {
                         if (!channelproperties.find((el) => { return val == el ;})) {
-                            console.log("WARNING : One property of the parameter of CallFactory method is not present in the Call class : ");
+                            //console.log("WARNING : One property of the parameter of CallFactory method is not present in the Call class : ");
                         }
                     });
             }