+++ src/lib/common/models/Contact.ts
@@ -1121,7 +1121,7 @@ class Contact {
                         return val == el;
                     })) {
                         //console.log("WARNING : One property of the parameter of BubbleFactory method is not present in the Bubble class : ", val, " -> ", data[val]);
-                        console.log("WARNING : One property of the parameter of obj method is not present in the Contact class : ", val);
+                        //console.log("WARNING : One property of the parameter of obj method is not present in the Contact class : ", val);
                     }
                 });
     }