+++ src/lib/common/models/Conversation.ts
@@ -342,7 +342,7 @@ class Conversation {
 
         //update last activity date for rooms when we receive/sent messages
         // dev-code //
-        console.log("conversation bubble : ", this.bubble);
+        //console.log("conversation bubble : ", this.bubble);
         // end-dev-code //
         if (this.bubble) {
             this.bubble.lastActivityDate = this.lastModification;